reworks client tests to be in line with other impls
